+++ src/main/java/nl/novi/backend_eindopdracht/controller/user/TrainerController.java
@@ -23,7 +23,7 @@ public class TrainerController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Trainer> getTrainerById(@PathVariable int id) {
+    public ResponseEntity<Trainer> getTrainerById(@PathVariable Long id) {
         return trainerService.getTrainerById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
@@ -36,7 +36,7 @@ public class TrainerController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteTrainer(@PathVariable int id) {
+    public ResponseEntity<Void> deleteTrainer(@PathVariable Long id) {
         trainerService.deleteTrainer(id);
         return ResponseEntity.noContent().build();
     }
